+++ hypervisor/include/arch/x86/page.h
@@ -77,8 +77,6 @@ struct pgtable {
 	void (*recover_exe_right)(uint64_t *entry);
 };
 
-void init_ept_pgtable(struct pgtable *table, uint16_t vm_id);
 struct page *alloc_page(struct page_pool *pool);
 void free_page(struct page_pool *pool, struct page *page);
-void reserve_buffer_for_ept_pages(void);
 #endif /* PAGE_H */
